requirements/requirements.txt
django>=1.6
south

django-tastypie
